+++ action/action-util/src/main/java/com/powsybl/action/util/Scalable.java
@@ -21,11 +21,11 @@ public interface Scalable {
 
     Logger LOGGER = LoggerFactory.getLogger(Scalable.class);
 
-    float initialValue(Network n);
+    double initialValue(Network n);
 
     void reset(Network n);
 
-    float maximumValue(Network n);
+    double maximumValue(Network n);
 
     void listGenerators(Network n, List<Generator> generators, List<String> notFoundGenerators);
 
@@ -33,7 +33,7 @@ public interface Scalable {
 
     List<Generator> listGenerators(Network n);
 
-    float scale(Network n, float asked);
+    double scale(Network n, double asked);
 
     static GeneratorScalable gen(String id) {
         return new GeneratorScalable(id);
